Refactoring, moving script functionality in under the package
